modularized commands and fixed timer errors
